refactor(units): split up units header (#12)
* refactor(units): create units folder (#4)

* refactor(units): SI header (#4)

* refactor(units): aggregate header (#4)

* refactor(units): SI length header (#4)

* refactor(units): SI time header (#4)

* format(units): format run

* fix(format): missing dependent files

* fix(units): missing include

* refactor(units): SI mass header (#4)

* refactor(units): SI electric current header (#4)

* refactor(units): SI temperature header (#4)

* refactor(units): SI substance amount header (#4)

* refactor(units): SI luminous intensity header (#4)

* refactor(units): substance amount tests (#4)

* refactor(units): SI time tests (#4)

* refactor(units): SI length tests (#4)

* refactor(units): SI mass tests (#4)

* refactor(format): SI electric current tests (#4)

* refactor(units): SI temperature tests (#4)

* refactor(units): SI luminous intensity tests (#4)

close #4
