+++ posu/units/si/ipp/luminous_intensity.ipp
@@ -1,0 +1,199 @@
+
+namespace posu::units {
+
+    inline namespace literals {
+
+        inline namespace luminous_intensity_literals {
+
+            [[nodiscard]] constexpr auto operator""_acd(unsigned long long value) -> attocandelas
+            {
+                return attocandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_acd(long double value)
+                -> luminous_intensity<double, std::atto>
+            {
+                return luminous_intensity<double, std::atto>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_fcd(unsigned long long value) -> femtocandelas
+            {
+                return femtocandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_fcd(long double value)
+                -> luminous_intensity<double, std::femto>
+            {
+                return luminous_intensity<double, std::femto>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_pcd(unsigned long long value) -> picocandelas
+            {
+                return picocandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_pcd(long double value)
+                -> luminous_intensity<double, std::pico>
+            {
+                return luminous_intensity<double, std::pico>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_ncd(unsigned long long value) -> nanocandelas
+            {
+                return nanocandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_ncd(long double value)
+                -> luminous_intensity<double, std::nano>
+            {
+                return luminous_intensity<double, std::nano>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_ucd(unsigned long long value) -> microcandelas
+            {
+                return microcandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_ucd(long double value)
+                -> luminous_intensity<double, std::micro>
+            {
+                return luminous_intensity<double, std::micro>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_mcd(unsigned long long value) -> millicandelas
+            {
+                return millicandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_mcd(long double value)
+                -> luminous_intensity<double, std::milli>
+            {
+                return luminous_intensity<double, std::milli>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_ccd(unsigned long long value) -> centicandelas
+            {
+                return centicandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_ccd(long double value)
+                -> luminous_intensity<double, std::centi>
+            {
+                return luminous_intensity<double, std::centi>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_dcd(unsigned long long value) -> decicandelas
+            {
+                return decicandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_dcd(long double value)
+                -> luminous_intensity<double, std::deci>
+            {
+                return luminous_intensity<double, std::deci>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_cd(unsigned long long value) -> candelas
+            {
+                return candelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_cd(long double value)
+                -> luminous_intensity<double, std::ratio<1>>
+            {
+                return luminous_intensity<double, std::ratio<1>>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_dacd(unsigned long long value) -> decacandelas
+            {
+                return decacandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_dacd(long double value)
+                -> luminous_intensity<double, std::deca>
+            {
+                return luminous_intensity<double, std::deca>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_hcd(unsigned long long value) -> hectocandelas
+            {
+                return hectocandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_hcd(long double value)
+                -> luminous_intensity<double, std::hecto>
+            {
+                return luminous_intensity<double, std::hecto>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_kcd(unsigned long long value) -> kilocandelas
+            {
+                return kilocandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_kcd(long double value)
+                -> luminous_intensity<double, std::kilo>
+            {
+                return luminous_intensity<double, std::kilo>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Mcd(unsigned long long value) -> megacandelas
+            {
+                return megacandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Mcd(long double value)
+                -> luminous_intensity<double, std::mega>
+            {
+                return luminous_intensity<double, std::mega>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Gcd(unsigned long long value) -> gigacandelas
+            {
+                return gigacandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Gcd(long double value)
+                -> luminous_intensity<double, std::giga>
+            {
+                return luminous_intensity<double, std::giga>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Tcd(unsigned long long value) -> teracandelas
+            {
+                return teracandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Tcd(long double value)
+                -> luminous_intensity<double, std::tera>
+            {
+                return luminous_intensity<double, std::tera>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Pcd(unsigned long long value) -> petacandelas
+            {
+                return petacandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Pcd(long double value)
+                -> luminous_intensity<double, std::peta>
+            {
+                return luminous_intensity<double, std::peta>(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Ecd(unsigned long long value) -> exacandelas
+            {
+                return exacandelas(value);
+            }
+
+            [[nodiscard]] constexpr auto operator""_Ecd(long double value)
+                -> luminous_intensity<double, std::exa>
+            {
+                return luminous_intensity<double, std::exa>(value);
+            }
+
+        } // namespace luminous_intensity_literals
+
+    } // namespace literals
+
+} // namespace posu::units
