+++ posu/units/units.hpp
@@ -1,0 +1,6 @@
+#ifndef POSU_UNITS_UNITS_HPP
+#define POSU_UNITS_UNITS_HPP
+
+#include "posu/units/si/si.hpp"
+
+#endif // #ifndef POSU_UNITS_UNITS_HPP
